fix bug with 64 bits
git-svn-id: svn+ssh://scm.gforge.inria.fr/svnroot/vplants/vplants/trunk/PlantGL@4286 ab253cce-29fb-0310-bb2f-979600cdbdeb

+++ src/cpp/plantgl/algo/grid/tile.h
@@ -114,8 +114,8 @@ class ALGO_API Tile {
 	std::string getName() const;
 
     /// Return id of \e self.
-    const unsigned int getId() const {
-	return (unsigned int)this;
+    const size_t getId() const {
+	return (size_t)this;
     }
 
     /// Return num of \e self.
